--- integration/src/main/java/de/reflectoring/jiraalerts/integration/spring/IntegrationConfiguration.java
@@ -1,10 +1,0 @@
-package de.reflectoring.jiraalerts.integration.spring;
-
-import de.reflectoring.jiraalerts.jiracomponent.spring.JiraComponentConfiguration;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.context.annotation.Import;
-
-@Configuration
-@Import(JiraComponentConfiguration.class)
-public class IntegrationConfiguration {
-}
