+++ integration/src/main/java/io/reflectoring/jiraalerts/integration/spring/IntegrationConfiguration.java
@@ -1,0 +1,10 @@
+package io.reflectoring.jiraalerts.integration.spring;
+
+import io.reflectoring.jiraalerts.jiracomponent.spring.JiraComponentConfiguration;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.context.annotation.Import;
+
+@Configuration
+@Import(JiraComponentConfiguration.class)
+public class IntegrationConfiguration {
+}
