--- jira-component/src/main/java/de/reflectoring/jiraalerts/jiracomponent/connection/persistence/JiraConnectionDataRepository.java
@@ -1,9 +1,0 @@
-package de.reflectoring.jiraalerts.jiracomponent.connection.persistence;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface JiraConnectionDataRepository extends JpaRepository<JiraConnectionData, Long>{
-
-}
