--- jira-component/src/main/java/de/reflectoring/jiraalerts/jiracomponent/spring/JiraComponentConfiguration.java
@@ -1,9 +1,0 @@
-package de.reflectoring.jiraalerts.jiracomponent.spring;
-
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-@ComponentScan(basePackages = {"de.reflectoring.jiraalerts.jiracomponent"})
-public class JiraComponentConfiguration {
-}
