+++ jira-component/src/main/java/io/reflectoring/jiraalerts/jiracomponent/connection/persistence/JiraConnectionDataRepository.java
@@ -1,0 +1,9 @@
+package io.reflectoring.jiraalerts.jiracomponent.connection.persistence;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface JiraConnectionDataRepository extends JpaRepository<JiraConnectionData, Long>{
+
+}
