+++ jira-component/src/main/resources/io/reflectoring/jiraalerts/jiracomponent/configuration/JiraUrlConfigurationPanel.properties
