+++ jira-component/src/test/resources/io/reflectoring/jiraalerts/jiracomponent/configuration/jira-configuration-test.properties
