+++ src/main/resources/io/reflectoring/jiraalerts/spring/hibernate.properties
